Revert "Simulations for Simple Model"
