+++ .gitignore
@@ -24,10 +24,3 @@ bld/
 *.pdf
 *.run.xml
 *.synctex.gz
-*.synctex.gz.sum.synctex
-*.toc
-*.snm
-*.nav
-
-# Jupyter notebooks
-*.ipynb
