--- paper/mte_papers.md
@@ -1,38 +1,0 @@
-# MTE Papers
-
-A collection of papers related to the MTE framework.
-In particular, applied papers that employ the framework of Mogstad et al. ECMA.
-
-## Applied Papers using Mogstad et al. ECMA Framework
-
-### Rose and Shem-Tov 2021 JPE
-
-Title: *How Does Incarceration Affect Reoffending? Estimating the Dose-Response Function*
-
-They use and extend the MTE model for extrapolation. They don't provide inference.
-For example, see Figure 5 and Table 6.
-
-### Koichiro, Takanori Makoto AER
-
-Title: Selection on Welfare Gains: Experimental Evidence from Electricity Plan Choice†
-
-They don't seem to primarily use the MST method, but rather the parametric method of Brinch et al.
-
-
-### Shea WP
-
-https://jkcshea.github.io/
-
-# To Check
-
-- https://direct.mit.edu/rest/article-abstract/doi/10.1162/rest_a_01483/124131/The-Value-of-Piped-Water-and-Sewers-Evidence-from?redirectedFrom=fulltext
-- https://direct.mit.edu/rest/article-abstract/105/3/646/102834/Reconciling-Seemingly-Contradictory-Results-from
-- https://academic.oup.com/ej/article-abstract/132/646/2231/6548192
-- https://academic.oup.com/restud/article-abstract/90/1/432/6582594
-- https://papers.ssrn.com/sol3/papers.cfm?abstract_id=4405043
-- https://www.aeaweb.org/articles?id=10.1257/aer.20181756
-- Appendix of https://academic.oup.com/qje/article-abstract/133/4/1885/5025665
-
-# Misc
-
-- Mismeasurement tratment: https://www.sciencedirect.com/science/article/pii/S0304407623002725
