+++ paper/refs.bib
@@ -115,25 +115,3 @@
   journal={arXiv preprint arXiv:2404.14603},
   year={2024}
 }
-
-@article{hong2018numerical,
-  title={The numerical delta method},
-  author={Hong, Han and Li, Jessie},
-  journal={Journal of Econometrics},
-  volume={206},
-  number={2},
-  pages={379--394},
-  year={2018},
-  publisher={Elsevier}
-}
-
-@article{fang2019inference,
-  title={Inference on directionally differentiable functions},
-  author={Fang, Zheng and Santos, Andres},
-  journal={The Review of Economic Studies},
-  volume={86},
-  number={1},
-  pages={377--412},
-  year={2019},
-  publisher={Oxford University Press}
-}
