+++ paper/thesis.tex
@@ -292,7 +292,7 @@ Note we would like to set $\theta_{30} = 0$ but this violates the constraint $\t
 	\caption{Upper Bound with Monotone MTR Functions}\label{fig:sm_upper_incr.tex}
 \end{figure}
 
-Figure~\ref{fig:sm_sol_upper} displays the solution to the upper bound $\overline{\beta^*}$ and lower bound $\underline{\beta^*}$ as a function of $\beta_s$ under different restrictions.
+Figure~\ref{fig:sm_sol_upper} displays the solution to the upper bound $\overline{\beta^*}$ and lower bound $\underleftarrow{\beta^*}$ as a function of $\beta_s$ under different restrictions.
 \begin{figure}
 	\centering
 	\begin{subfigure}{0.45\textwidth}
@@ -335,76 +335,6 @@ Using the simplification noted above, the estimator can be equivalently written 
 	.
 \end{equation}
 
-\section{Inference Approaches}
-
-The target is to make inference on the parameter $\beta^*$ which for the simulation purposes I set to $LATE(0.4, 0.6)$.
-\footnote{Inference for the interval-identified parameter $\beta^*$ and its identified set are not necessarily the same as pointed out by \cite{imbens2004confidence} in simple examples. I will return to this point below.}
-That means, we want to extrapolte the LATE for the instrument-complier subpopulation with $u\in[0.4, 0.6]$ to a broader sub-population.
-Inference is complicated by the fact that the optimal solution $\beta^* = \beta^*(\beta_s)$ exhibits a kink at $\beta_s=0$ and hence is not fully differentiable at that point.
-However, as pointed out by~\cite{fang2023inference} and~\cite{hong2018numerical}, a delta method can be used that only requires \textit{directional} differentiability of the functional.
-~\cite{fang2023inference} propose a method that combines a bootstrap approximation to the underlying parameter with an estimate of the directional derivative of the functional $\phi$ that exploits its analytical structure.
-~\cite{hong2018numerical} on the other hand propose to use a numerical approximation to the directional derivative.
-Henceforth, I will call these methods the \textit{analytical} and \textit{numerical} delta bootstrap, respectively.
-
-\subsection*{Notation}
-In the following I will use the notation of~\cite{fang2023inference}.
-Our goal is to perform inference on the real-valued parameter $\beta^*$, which is only set-identified.
-As shown above the set is given by $[\underline{\beta^*}(\beta_s), \overline{\beta^*}]$,
-with the following explicit solutions:
-\begin{equation*}
-	\overline{\beta^*}(\beta_s)=
-	\begin{cases}
-		\omega \beta_s + (1 - \omega),& \quad \text{if } \beta_s \geq 0\\
-		\beta_s + (1 - \omega),              & \quad \text{if } \beta_s < 0,
-	\end{cases}
-\end{equation*}
-and
-\begin{equation*}
-	\underline{\beta^*}(\beta_s)=
-	\begin{cases}
-		\beta_s - (1 - \omega),& \quad \text{if } \beta_s \geq 0\\
-		\omega \beta_s - (1 - \omega),              & \quad \text{if } \beta_s < 0.
-	\end{cases}
-\end{equation*}
-In the notation of~\cite{fang2023inference} we have an underlying parameter $\theta_0 \equiv \beta_0^s$ which in our case is real-valued.
-The mapping $\phi: \mathbb{R} \to \mathbb{R}$ is the solution to the upper or lower bound, respectively.
-That is, we have $\overline{\phi} \equiv \overline{\beta^*}$ with the mapping given above.
-As is evident, this function has a kink at $\beta_s = 0$ and hence is only directionally differentiable at that point.
-
-If we can find a method to construct confidence intervals for $\overline{\phi}(\beta_0^s)$ and $\underline{\phi}(\beta_0^s)$, we can also construct a confidence interval.
-
-
-Note that our setting is the same as Example 2.2 of~\cite{hong2018numerical}, which itself is a more general version of Example 2 in~\cite{fang2023inference}.
-
-\section{Simulation Results}
-
-\begin{figure}
-	\centering
-	\begin{subfigure}{0.45\textwidth}
-		\centering
-		\includegraphics[width=\textwidth]{../bld/boot/figures/coverage.png}
-		\caption{Coverage for True Parameter}
-	\end{subfigure}
-	\hfill
-	\begin{subfigure}{0.45\textwidth}
-		\centering
-		\includegraphics[width=\textwidth]{../bld/boot/figures/length.png}
-		\caption{Length of CIs}
-	\end{subfigure}
-
-	\begin{subfigure}{0.45\textwidth}
-		\centering
-		\includegraphics[width=\textwidth]{../bld/boot/figures/means.png}
-		\caption{Means of Upper and Lower Bound}
-	\end{subfigure}
-	\hfill
-	\begin{subfigure}{0.45\textwidth}
-		\centering
-		\includegraphics[width=\textwidth]{../bld/boot/figures/coverage_eps_fun.png}
-		\caption{Coverage by Step Size for Numerical Delta Method}
-	\end{subfigure}
-	\caption{Simulation: CI Properties for Standard and Numerical Delta Bootstrap}\label{fig:simulation_ci_bootstrap}
-\end{figure}
 
 \bibliographystyle{plain}
 \bibliography{refs.bib}
