+++ pyproject.toml
@@ -12,9 +12,7 @@ pdbcls = "pdbp:Pdb"
 target-version = "py311"
 select = ["ALL"]
 fix = true
-# Might want to reduce this at some point, currently just the set to whatever number
-# is required.
-pylint.max-args = 13
+pylint.max-args = 9
 fix-only = false # No linting errors will be reported
 extend-ignore = [
     "S101",  # Use of `assert` detected.
