+++ src/thesis/classes.py
@@ -1,5 +1,4 @@
 """Custom classes for thesis."""
-
 from collections.abc import Callable
 from dataclasses import dataclass
 from typing import NamedTuple
@@ -72,11 +71,3 @@ class MonteCarloSetup(NamedTuple):
     sample_size: int
     repetitions: int
     u_hi_range: np.ndarray | None = None
-
-
-class LocalATEs(NamedTuple):
-    """LATEs for subpopulations: never-taker, complier, always-taker."""
-
-    always_taker: float
-    complier: float
-    never_taker: float
