--- src/thesis/misc/__init__.py
@@ -1,1 +1,0 @@
-"""Miscaellaneaous scripts and notebooks for the thesis."""
