--- src/thesis/r_ivmte/.Rhistory
@@ -1,96 +1,0 @@
-library("AER")
-library("ivmte")
-knitr::kable(head(AE, n = 10))
-lm(data = AE, worked ~ morekids)
-lm(data = AE, morekids ~ samesex)
-ivreg(data = AE, worked ~ morekids | samesex )
-# Simple bounds for ATT using moment approach
-results <- ivmte(data = AE,
-target = "att",
-m0 = ~ u + yob,
-m1 = ~ u + yob,
-ivlike = worked ~ morekids + samesex + morekids*samesex,
-propensity = morekids ~ samesex + yob,
-noisy = TRUE)
-# Simple bounds for ATT using moment approach
-results <- ivmte(data = AE,
-target = "att",
-m0 = ~ u + yob,
-m1 = ~ u + yob,
-ivlike = worked ~ morekids + samesex + morekids*samesex,
-propensity = morekids ~ samesex + yob,
-noisy = TRUE)
-library("AER")
-library("ivmte")
-# install.packages("ivmte")
-# install.packages("splines2 ")
-install.packages("lpSolveAPI ")
-library("AER")
-library("ivmte")
-knitr::kable(head(AE, n = 10))
-lm(data = AE, worked ~ morekids)
-lm(data = AE, morekids ~ samesex)
-ivreg(data = AE, worked ~ morekids | samesex )
-# Simple bounds for ATT using moment approach
-results <- ivmte(data = AE,
-target = "att",
-m0 = ~ u + yob,
-m1 = ~ u + yob,
-ivlike = worked ~ morekids + samesex + morekids*samesex,
-propensity = morekids ~ samesex + yob,
-noisy = TRUE)
-library(lpSolveAPI)
-# install.packages("ivmte")
-# install.packages("splines2 ")
-install.packages("lpSolveAPI ")
-# install.packages("ivmte")
-# install.packages("splines2 ")
-install.packages("lpSolveAPI")
-# install.packages("ivmte")
-install.packages("splines2")
-library("AER")
-library("ivmte")
-knitr::kable(head(AE, n = 10))
-lm(data = AE, worked ~ morekids)
-lm(data = AE, morekids ~ samesex)
-ivreg(data = AE, worked ~ morekids | samesex )
-# Simple bounds for ATT using moment approach
-results <- ivmte(data = AE,
-target = "att",
-m0 = ~ u + yob,
-m1 = ~ u + yob,
-ivlike = worked ~ morekids + samesex + morekids*samesex,
-propensity = morekids ~ samesex + yob,
-noisy = TRUE)
-library(lpSolveAPI)
-# Simple bounds for ATT using moment approach
-results <- ivmte(data = AE,
-target = "att",
-m0 = ~ u + yob,
-m1 = ~ u + yob,
-ivlike = worked ~ morekids + samesex + morekids*samesex,
-propensity = morekids ~ samesex + yob,
-noisy = TRUE)
-# Parametric restrictions on m0, m1
-args <- list(data = AE,
-ivlike =  worked ~ morekids + samesex + morekids*samesex,
-target = "att",
-m0 = ~ u + I(u^2) + yob + u*yob,
-m1 = ~ u + I(u^2) + I(u^3) + yob + u*yob,
-propensity = morekids ~ samesex + yob)
-r <- do.call(ivmte, args)
-r
-AE
-length(AE)
-len(AE)
-shape(AE)
-a = AE
-# Parametric restrictions on m0, m1
-args <- list(data = AE,
-ivlike =  worked ~ morekids + samesex + morekids*samesex,
-target = "att",
-m0 = ~ u + I(u^2) + yob + u*yob,
-m1 = ~ u + I(u^2) + I(u^3) + yob + u*yob,
-propensity = morekids ~ samesex + yob)
-r <- do.call(ivmte, args)
-r
