--- src/thesis/r_ivmte/__init__.py
@@ -1,1 +1,0 @@
-"""Explore R ivmte package: https://github.com/jkcshea/ivmte/."""
