+++ tests/bootstrap_fail/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for the bootstrap_fail functions."""
