+++ tests/bootstrap_fail/test_funcs.py
@@ -1,0 +1,56 @@
+"""Test functions for bootstrap_fail."""
+
+import numpy as np
+import pandas as pd  # type: ignore[import-untyped]
+import pytest
+from thesis.bootstrap_fail.funcs import _draw_data, _estimate_pscores
+from thesis.classes import Instrument
+from thesis.config import RNG
+
+
+@pytest.fixture()
+def instrument() -> Instrument:
+    return Instrument(
+        support=np.array([0, 1]),
+        pmf=np.array([0.5, 0.5]),
+        pscores=np.array([0.4, 0.6]),
+    )
+
+
+def test_data_moments_boundary(instrument) -> None:
+    n_obs = 100_000
+
+    expected = pd.DataFrame(
+        {
+            "y_given_z": [0.0, 0.0],
+            "d_given_z": instrument.pscores,
+        },
+        index=[0, 1],
+    )
+
+    data = pd.DataFrame(
+        _draw_data(n_obs, RNG, instrument=instrument, param_pos="boundary"),
+        columns=["y", "d", "z", "u"],
+    )
+
+    actual = pd.DataFrame(
+        {
+            "y_given_z": data.groupby("z")["y"].mean(),
+            "d_given_z": data.groupby("z")["d"].mean(),
+        },
+        index=[0, 1],
+    )
+
+    pd.testing.assert_frame_equal(actual, expected, atol=0.01)
+
+
+def test_compute_pscores(instrument) -> None:
+    n_obs = 1_000_000
+
+    data = _draw_data(n_obs, RNG, instrument=instrument, param_pos="boundary")
+
+    expected = (0.4, 0.6)
+
+    actual = _estimate_pscores(data)
+
+    assert expected == pytest.approx(actual, abs=3 / np.sqrt(n_obs))
