+++ tests/imbens_manski/test_funcs.py
