--- tests/simple_model/__init__.py
@@ -1,1 +1,0 @@
-"""Tests for the simple_model functions."""
