--- tests/utilities/__init__.py
@@ -1,1 +1,0 @@
-"""Tests for utilities."""
